chore: move get_url to test_unit
as its not used in jobs
also
- cleaned up imports
- remove spatial file debugging artifact that was used during devt
- expanded description of jinja2 formulae processing
